Update readme (#166)
* Update readme

More cleanup of the readme

Alphabetized lists of complete and WIP items, added a link to the todo for modules, add links to our Travis and AppVeyor CIs in contributing, and adjust a few language items.

* Undo an indent

Didn't need to indent one instruction in installation.

* Unindent all of those...

Unindented all of the instructions for installation and running. Gah.