fix timezone issues
the open/close times were being interpreted with the current timezone's
dayofyear, but the target timezone's timeofday. Which is correct, I
guess, since we don't give it a day to work with.

This takes the day of the year from CST and uses it as the base for the
open/close times.