Add iOS notifications
Using Cocoapods to add the code driving the OneSignal service for the notificaitons. The backend is registering the simulator that I'm using for iOS testing, but since simulators cannot get notificaitons I haven't seen the end-to-end process work yet. I'll be able to do that within the day and will make an update if it's not working as expected.