Add iOS notifications (#161)
* Remove ride-along change

Didn't intend for that to be in here, was just faster to make another commit than to revert.

* Use relative paths

Xcode used absolute paths. I want to use relative paths. I am smarter than Xcode.

* Make Travis work with Cocoapods (#164)

* Change the test command to use the workspace file

* switch to xcodebuild

* Revert "switch to xcodebuild"

fc5fa3e8fc3e852598227be56ebdabc07866b10b

* switch to just building iOS

* Add iOS notifications

Using Cocoapods to add the code driving the OneSignal service for the notificaitons. The backend is registering the simulator that I'm using for iOS testing, but since simulators cannot get notificaitons I haven't seen the end-to-end process work yet. I'll be able to do that within the day and will make an update if it's not working as expected.

* missed a thing while rebasing