Calendar update (#305)
* Remove unnessary functions

I like the new pattern of putting the arrow functions in the listviews, so I chaged it to use that.

* Re-format dates so that we are not showing past days at the top of the calendar

The old system used the start day of multi-day events to display the event, even if that day had passed. It looked really silly to have "Sept. 7" at the top of the calendar still, so I force it to use today.

* Add missing braces to statement

Missing braces to "if" statement fixed.