update travis config and steps
- adds ios builds back to travis
- allows automatic once-daily betas from travis (via cron)
- uses fastlane to build both ios and android
- adds a js-bundling test to the js builds
- only pings coveralls after the js build

Some more info: I realized that the caches, and `before_cache`
instructions, didn't need to be different per-platform – node_modules
should be cached everywhere, for instance.

That allowed me to get rid of the yaml references.

I also upgraded the js builds to the new "trusty" beta. Not much
difference there. The android builds are still on "precise", because the
android stuff isn't available on trusty yet.

I converted the env configs into objects, to let travis turn them into
the strings for us, and I renamed some of them, preferring lowercase keys.

OS X has been upgraded to Sierra and Xcode 8.1.

Travis will now spawn a cron-job to trigger a build once a day; this
build will release a new beta.

I've started exporting the api key into the build, if we're going to
deploy.

I added a JS bundle step to the js builds.

That's about it.