Add iOS beta-deployment abilities to Travis (#414)
* add `fastlane ios update_match`

to forcibly run match and update the provisioning profiles

* consolidate the keychain setup into a fastlane lane

this is just simpler than letting it spread out over everything.

* make `fastlane ios build` available

because travis is still failing to run the UI tests

fastlane readme

* add `fastlane ios screenshot` command

also updates the Snapfile to put the screenshots in the .gitignored place

* fix operator precedence issue in fastfile

* encode some build information in the changelog

add the branch and hash to the changelog.

also updates the changelog to report all commits since last build.

* change match repo url to https

Travis and the Deploy Key approach recommend (require?) the use of https
instead of git+ssh.

* fix android increment_version_code action path

increment_version_code uses a glob to look for the build.gradle file,
but it takes the first one it finds - in a pattern like
'**/$app_folder_name/*/build.gradle'

So we're doing two things here:
- restricting the glob to the build.gradle we want,
- and removing the "./" from the pattern so it will match the paths

* ignore ruby deps from android builds

* update fastlane

* fix hockeysdk installation script

* remove bitrise config

* update travis config and steps

- adds ios builds back to travis
- allows automatic once-daily betas from travis (via cron)
- uses fastlane to build both ios and android
- adds a js-bundling test to the js builds
- only pings coveralls after the js build

Some more info: I realized that the caches, and `before_cache`
instructions, didn't need to be different per-platform – node_modules
should be cached everywhere, for instance.

That allowed me to get rid of the yaml references.

I also upgraded the js builds to the new "trusty" beta. Not much
difference there. The android builds are still on "precise", because the
android stuff isn't available on trusty yet.

I converted the env configs into objects, to let travis turn them into
the strings for us, and I renamed some of them, preferring lowercase keys.

OS X has been upgraded to Sierra and Xcode 8.1.

Travis will now spawn a cron-job to trigger a build once a day; this
build will release a new beta.

I've started exporting the api key into the build, if we're going to
deploy.

I added a JS bundle step to the js builds.

That's about it.