<FilterView/> component (#467)
* Add a FilterView component

Usage:
<FilterView
  // a path to the filters stored in Redux somewhere
  pathToFilters={['menus', 'stav']}
  // you need to do something with the changes
  onChange={filters => doSomething()}
/>

In React parlance, FilterView is a "controlled component" – it will only
ever render what it is given, and it does not maintain any state itself.

It can render two types of filter, a Toggle and a List. Types for those
are provided in `types.js`.

* Add an `applyFiltersToItem` function

Usage: arr.filter(item => applyFiltersToItem(filters, item))

Applies all filters in `filters` to `item`. Returns `true` if they pass,
and `false` if any fail.