fix indentation of building hour submissions
"dedent" works by removing all comon leading whitespace in a
block of text, so when I unindented the closing backtick by two
spaces, it left the rest of the block prefixed by two spaces.

I also removed the indentation from the email body, and
attempted to use html in the body, as well; we'll see if that
worked.