rework the propagate_version action
It now takes the place of the set_version action, as well.

It takes the version from package.json and writes it into the gradle and
Xcode config files, thereby allowing us to never need to update those
version numbers.

This will allow us to release a new version entirely from GitHub, as
we'll be able to just update the version in package.json and make a new
release, without needing any command-line tools.