rework BusView's internals
Our data is formatted for humans, which means that the stops are defined separately from the stop _times_, so that you can look down the arrays to line them up visually.

However, this view works on a stop-by-stop basis.

Before, what we'd do is pick out the times that went with each stop as we rendered that stop.

With this patch, the first thing we do is to process the two arrays into a "timetable" array of objects that contains all the data needed to render the rows.

This colocation helps to reduce the amount of data we need to pass around and recalculate, and I think that it really simplifies the internals of this view.