switch to functional array manipulation in groupBuildings
We avoid manual array manipulation in the app, in favor of the .map/.reduce functions.

You can see how the first loop, which builds up an array of favorited buildings, can be
replaced with a single line Array#filter call that "filters" down the existing array to just
the items we're interested in. I also merged the array creation with the group creation on
the previous line.

Then, later, we had some array manipulation to build up the total set of groups to show; I
replaced the if-and-loop with a single if that contains an "array spread" to prepend the
Favorites group if we need to.