remove SAVE_FAVORITE_BUILDINGS action in favor of TOGGLE_FAVORITE_BUILDING and LOAD_FAVORITE_BUILDINGS
LOAD_FAVORITE_BUILDINGS is a single-purpose action that loads the favorites from AsyncStorage
and tells the reducer to store them.

TOGGLE_FAVORITE_BUILDING is a much fancier action creator, that gets the current set of
favorited buildings from the redux state and uses that to add or remove the given building
name from the set of favorites, and then re-sort the list of buildings.

It also has to save the favorites to AsyncStorage, because there's no better place to perform
the actual "save" (yet).