use React.PureComponent in CourseResultsList
I've been removing PureComponent as I wander because, as [1] points
out, "If a component usually changes when there’s an update, then a
PureComponent will be doing two diffs instead of just one (props and
state in shouldComponentUpdate, and then the normal element diff). Which
means it’s going to be slower usually but faster occasionally."

PureComponent's not _bad_, but it's not a panacea either. It's really,
really useful in some situations, though. Like here.

[1]: https://cdb.reacttraining.com/react-inline-functions-and-performance-bdff784f5578#f11e