use a [key] to automatically recreate the list when searching
This avoids the need to scroll back to the top manually, because the
initial state for a list is to be at the top. Because the search cache
is at the module level, and not cached within an instance of the
component, we don't incur a speed penalty for recreating the search
list either.

> In order to reset the value when moving to a different item […], we
> can use the special React attribute called key. When a key changes,
> React will create a new component instance rather than update the
> current one

https://reactjs.org/blog/2018/06/07/you-probably-dont-need-derived-state.html#recommendation-fully-uncontrolled-component-with-a-key