fix padding on Credits view
ScrollView's an odd beastie. My mental model is kinda that you have two
views, an outer and an inner, and [style] applies to the outer while
[contentContainerStyle] applies to the inner.

And the only one that actually scrolls is the inner one.

If you apply a backgroundColor to the inner, you can see that the color
pulls away from the edge of the screen as it follows the rubber-banding
list, as opposed to applying it to the outer view, where it never moves
from the edge of the screen.

Padding, though, and margin, … it's like … adding padding to the outer
one just pushes down the starting point of the inner one? It's like the
padding is only applied to the top, and not to the bottom, when
computing where the "scrolling" container should be.

TL;DR: put backgroundColor on [style], and padding/margin on [contentContainerStyle].