gitattributes: Use merge=union on CHANGELOG.md
This tells Git to always use union generation as its preferred way of
resolving merge conflicts.

Whereas git-merge-file(1) typically will ordinarily produce merge
conflicts, this explicitly tells Git that we would prefer that this
file include the union of all changes proposed.

Hence, if someone adds

  # Added
  - Added a feature

and their PR gets merged to master but not to our branch, and we have:

  # Added
  - Added a different feature

whereas `git merge-file -p` reports

  # Added
  <<<<<<< CHANGELOG.master
  - Added a feature
  =======
  - Added a different feature
  >>>>>>> CHANGELOG.different-branch

in this case, `git merge-file -p --union` with the same inputs produces:

  # Added
  - Added a feature
  - Added a different feature

While this is an unverified claim, I think that in production Git will
be following DAG-order when it makes these decisions.  Hence, a merge
into master of a branch that does not have an up-to-date changelog
will append that changelog's contents to the relevant section in the
upstream, thereby preserving both parts of history.  If a merge is
possible, Git will successfully run that.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>