fastlane: Remove unnecessary Encoding comments
At RuboCop's behest, I removed the "coding: utf-8" comments at the top
of our Ruby files.  Emacs was adding these automatically for me,
because of the `ruby-insert-encoding-magic-comment' variable being
truthy.  I have set a hook to disable it now.

It's worth noting that I am still not a fan of using Unicode in source
code because it's better to just use ASCII characters that we can
_guarantee_ will work.  But this is okay.

I verified that this change was complete via:

  bundle exec rubocop --only 'Style/Encoding'

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>