Explicitly build nightlies
Since ios-nightly and android-nightly are built off of the ios and
android builds, I don't think we actually have any safeguards to
ensure that we do actually build nightlies besides setting environment
variables.

We should explicitly state that we want nightlies to be built.

This commit does that; if `ENV['IS_NIGHTLY']` is anything besides nil
or false, it will report that the build should continue as planned.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>