ssb: Extract nightly build logic into its own function
In case we want to add anything else.

This is done to make the first line of should_build? a bit more
idiomatic.  I read this now as "return [whether or not I should build
the nightly] if [the nightly environment is present]"

Blah blah blah.

Stack depth doesn't matter.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>