parseHours: Handle dayOfYear wraparound
I really don't know if this is anywhere near correct, but it seems to
resolve the Jest failures we're getting.

Previously, `m`'s dayOfYear was applied to both `open` and `close`, but
if `m` was in the next year, then its `dayOfYear` would be low, and
 `open` and `close` would be freely wrapped around to the beginning of
the year.

Now, `parseHours` checks if `open` or `close` are six months prior to
`m`, and if so, adds `1` year.  `m` is used only to provide context to
"10am" and such kinds of strings, so I think this is a safe bet.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>