parseTime: Handle dayOfYear wraparound
Previously, `now`'s `dayOfYear` would be applied to the parsed time,
and when `now` and the parsed time were in different years the
`dayOfYear` from the parsed `now` would be applied, thereby wrapping
the parsed date back around to the beginning of the year.

This commit checks if our parsed date is more than six months before
`now`, and if so, it adds a year to it.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>