Remove react-native-google-analytics-bridge
Analytics is not helpful in our use case, and presents a possible
privacy concern for users.  We have decided to at least remove
analytics for the time being, especially considering that Google
Analytics for mobile apps will be sunset in October of this year.

We do not want to move towards "Firebase" or other Google-things,
instead opting to add analytics back selectively where we think
information may be useful.

As we move towards other crash-tracking systems such as Sentry, we can
also investigate whether it is sufficient to send copious
breadcrumbs.  Otherwise, we will implement our own analytics system,
so we can tune our retention and privacy policies accordingly.

This commit does a large set of things, but each of the changes is
required to keep functionality.

- Remove `react-native-google-analytics-bridge` from our dependency
  manifests and build configurations.

- Remove redux wiring for "Feedback" (which is Analytics, I believe.)

- Remove google-analytics from our modules/analytics thing, and remove
  all references to its exports.

I have working flow, jest, and prettier results, so I am pushing for
initial review.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>