add "specs" key as required for detox upgrade
https://github.com/wix/Detox/blob/master/docs/Guide.Migration.md#migrating-from-detox-120x-to-121x

For the curious ones, who want to know more why we should use an empty string ("") instead of
deleting "specs" and "file" from package.json, here is the explanation. This seemingly weird
step is motivated by backward compatibility with the previous versions of Detox.

So far, before detox@12.1.0, absence of file and specs properties implied a default test folder
value ("e2e"). In other words:

```
const testFolder = config.file || config.specs || "e2e";
```

In order not to break the existing logic but to introduce the deprecation, the check for the e2e
placeholder assignment became narrower yet remaining valid:

```
let testFolder = config.file || config.specs;
if (testFolder == null) { // that's why you should change it to an empty string, ""
    testFolder = "e2e"; // otherwise, if it is null or undefined, then we save backward compatibility
}
if (testFolder) { printDeprecationWarning(); }
```

As it can be seen above, this move allows to track if you followed the migration guide or not.