fastlane: Decentralize version code generation, fix Sentry sourcemaps
We were previously just sending the value of current_bundle_version as
both the RELEASE_NAME and DISTRIBUTION_NAME parameters for the Sentry
CLI, however, this was incorrect.  It seems that Sentry instead
associates with the provided DISTRIBUTION_NAME, (i.e. iOS'
CFBundleVersion or Android's versionCode)

So we were uploading with `2.8.0-pre` in both RELEASE_NAME and
DISTRIBUTION_NAME instead of `2.8.0-pre` for RELEASE_NAME and the
distribution version code for DISTRIBUTION_NAME

Now, we have a helper that can query against the gradle stuff and get
the appropriate values out.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>