Poking at the revert-revert of inline-requires
Context: Enabling inline requires might take more setup.
The initial blogpost about inline requires made it seem
as if all that was needed  was to flip this switch but
we've since learned that we'll crash if we don't do
_something_ else in the setup instructions. What that
is, I'm unsure.

Blog post and setup instructions:
* http://facebook.github.io/react-native/blog/2019/03/12/releasing-react-native-059
* https://facebook.github.io/react-native/docs/performance#ram-bundles-inline-requires

Related enabling and disabling of this setting:
* https://github.com/StoDevX/AAO-React-Native/pull/3573
* https://github.com/StoDevX/AAO-React-Native/pull/3666