html-lib: Update usage of htmlparser2
Inspired by [this line][a] and [this line][b], I decided to try simply
adding a 2.  Flow, Jest, and the rest of the goon squad seem to agree
that this looks like a good change.

Have not verified that this actually fixes the broken views, but for now
I will say that this seems to resolve #3921.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>

[a]: https://github.com/fb55/htmlparser2/commit/759b1220c03e55a895f971deab9f1e94c30a7f61#diff-04c6e90faac2675aa89e2176d2eec7d8L19
[b]: https://github.com/fb55/htmlparser2/commit/759b1220c03e55a895f971deab9f1e94c30a7f61#diff-04c6e90faac2675aa89e2176d2eec7d8R21