codecov: No longer produce commit statuses
We have found that code coverage is not generated for Detox testing,
which is (unsurprisingly) a _very_ big hole in our potential covered
codebase.

Core business logic should certainly be tested, but we have no real way
of measuring the coverage of our other areas of logic; this commit
suspends the red-herring alerts that cause _entire PRs_ to be visibly
"failing," despite being perfectly valid and well-tested.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>