fix time tests once and for all?
the core of this change is to pin the time tests to a specific date; I
chose `2019-12-18T18:39:45-0600`, because it is a recent date when all
of the tests passed.

I then set the moments created by our moment helper functions to the
year/month/day of the "base moment" [^1].

I followed up by removing the ±6-month logic from
https://github.com/StoDevX/AAO-React-Native/issues/3375; it seems to not
be needed anymore.

[^1]: I'm setting y/m/d instead of just dayOfYear because I need to set
      year as well. I could probably do just year().dayOfYear(), but I
      feel that this requires less thinking when you see it again.