ci: Attribute data deployment commits to github-actions[bot]
Using GitHub's users.noreply.github.com email addresses, which are of
the form <id+username@users.noreply.github.com>, we can attribute data
deployment commits to GitHub Actions.  I used the API response for the
user github-actions[bot] (which can be read at
https://api.github.com/users/github-actions%5Bbot%5D) to find the id
(41898282) and login ("github-actions[bot]"), which are then filled into
the form above to get the email address.  This has the benefit of
setting the avatar to something more official-looking, too.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>