stav-covid: Create an empty script for generating hours
In response to (GH) #4841, I decided to take a stab at parsing the hours
that are listed on the reslife site.

In the past, we have shied away from such automation because it's quite
hard to discern the proper source of truth.  BonApp has been known to do
quite frankly whatever they want, and the website has not had such a
table published.

In this case, however, Stav Hall now has hours published on a page. [1]
This is primarily because each dorm has a strict window within which
they are expected to get each meal.  These hours are (seemingly) updated
daily, which would produce an unreasonable maintenance burden.  This
script will be responsible for dumping the hours from the dining-hours
website into the 1-3-stav.yaml source file, which can then be parsed and
loaded client-side.

[1]: https://web.archive.org/web/*/https://wp.stolaf.edu/reslife/dining-hours/

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>