eslint: Stick with default --max-warnings
While it is nice sometimes to have all warnings be show-stoppers, in our
case right now it's better to just have the errors be show-stoppers.  As
long as the code can _compile_ (meaning it is at least mostly valid in
the eyes of the TS compiler) we can move forward.  We can look at being
more strict and forcing things to be errors later.  (E.g. via the config
file.)

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>