Remove proxied directory searching in favor of on-device
We lose results that are provided on-campus if insert ourselves as the
endpoint, and that unfortunately means this endpoint is a little more
fragile while being potentially receiving a lot of traffic. The tradeoff
to show student results seems right, though.

This also moves the formatting of the data into helpers and out of the
proxied server.