remove deprecated setEventSentSuccessfully callback
Rather than opt-in to sentry's beforeSend hook to capture an event id
and store it somewhere (provider, localstorage), we will most likely
be able to identify the new event id on the dashboard when we call this
method outside of development within the app. This allows us to avoid
adding more branching or complexity for the time being.

If we do want to go the path of capturing this event id, we can look at
adding the beforeSend hook to get the last-sent-event-id by calling
Sentry.getCurrentHub().lastEventId(). See this issue for more context.
https://github.com/getsentry/sentry-react-native/issues/787