progress on having fancy menu adopt the useAsync data fetching
There are type inconsistencies that are fixed with commented out code,
but the rendering logic breaks in this case.

If replaced with the commented-out version, we render the "Something
went wrong" path at line 266. If we wrap the grouping variables in a
useEffect (specialMessage, foodItems, and meals) and capture their
outputs via useState at the top of the page, we end up with undefined
data...