add username from login credentials instead of from navigation params
this is the same issue and fix that Hannes fix three years ago in a
separate PR that fixed releaseJob, whereas this fixes cancelJob
https://github.com/StoDevX/AAO-React-Native/pull/3730

we we were trying to extract the username from the navigation params
rather than loading them using loadLoginCredentials. I guess this never
got updated when we moved away from storing login credentials in redux
itself.