add support for overriding the close screen button title
Sometimes 'Done' sounds like a confirmation when instead we mean to say
this is a destructive action, like 'Close'.

Now with modals in the app we may want to override this title for
screens like the hours editor where it would be unfortunate if 'Done'
was interpreted as an action that saved what you were editing.