fix android compilation error caused by upstream alpha version
The 1.+ logic in react-native-inappbrowser-reborn's build.gradle allows
for upstream releases to be pulled in. It allowed the 1.5.* alpha of
androidx.browser.browser from google, released 10/24/2022, to break our
app.

This requires android sdk 33 which we are not set up to support. We need
to override the property in our own project so android builds
with 1.4.

https://github.com/proyecto26/react-native-inappbrowser/issues/386