explicitly check for a null return type when evaluating saved booleans
When fetching the saved value for getItemAsBoolean, we failed to fall
back to the correct value due to checking if the value was defined via
a check for "saved or default". Since these values are booleans, it is
problematic to check for a non-null value this way.

This became evident in two cases: first, the app's UI was reflecting a
falsey state the first run through. Next, if we overrode the default
value in the call, we would receive an incorrect UI toggle state.