Post adding support for Xcode 14 and macOS Ventura in #6573, the iOS build pipeline struggles to produce a build and dies in the last bit of the fastlane pipeline.
It looks like the suggested solution is to turn off signing for resource
bundles. An official fix is provided in release v0.70.3 of react native
in commit 967de03f304404ac8817936da37ca39514a09e33.

We can adopt a version of it from the CocoaPods repo into our Podfile and
should be able to remove this once Cocoapods catches up.
https://github.com/CocoaPods/CocoaPods/issues/11402#issuecomment-1201464693