tsc-counts: Match on any number of digits following TS prefix
With the upgrade to TypeScript v4.9.3, a new 5-digit error is generated
(TS18046) which gets snipped by the old \d{4} regex component which does
not include any word-boundary detection.  This leads to a loss of
information when trying to generate the list of TS error codes, and
ultimately programmer confusion.

This commit resolves GH-6592 by allowing any number of digits to follow
the TS prefix.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>