tsc-errors: Include word-boundary markers
Should TypeScript errors ever expand to include more than just digits in
the future, the way to detect the boundary of a typescript error code
will likely be the use of the \b word-boundary character. These ought to
be harmless at this point so I'm including them for didactic purposes
more than anything.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>