lock androidx annotation to 1.5.0 due to inappbrowser-reborn (again)
react-native-inappbrowser-reborn installs androidx.annotations@1.6.0 which has
a dependency on kotlin-stdlib:1.8.0. By default, gradle is using the higher
version of this lib in the project which causes incompatabilities between
kotlin-stdlib-jdk8 and kotlin-stdlib-jdk7. This is due to the merge of jdk7 &
jdk8 artifacts that occured in kotlin-stdlib:1.8.0, resulting in an an error if
you use packages that have kotlin-stdlib-jdk7/8 < 1.8.0. The latest includes a
kotlin 1.8.0 dependency that enables this issue and fails to build.