perf: use babel to remove console logs in production
When running a bundled copy of the app, console statements can cause a
big bottleneck in the JavaScript thread. This will automatically remove
all console.* calls in the release versions of this project. It is
recommended to use the plugin even if no console.* calls are made as a
third party library could also call them.