babel: Use plugin-transform-export-namespace-from
With ES2020, `export * as Namespace from 'source'` was merged into the
ECMAScript standard, and @babel/plugin-proposal-export-namespace-from
became deprecated.  The package, however, remained in our package-lock,
until Renovate removed it with lockfile maintenance.

The error message emitted when this line is removed suggests that we
simply need to reference `@babel/plugin-transform-export-namespace-from`
which differs only in that the name is "plugin-transform" instead of
"plugin-proposal".

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>